#64 paper-toast does not fire closed events
Attach listener to the `transitioned` event and call `_finishRenderOpened` or `_finishRenderClosed` respectively, causing the appropriate events to be fired on closing and opening the toast.
there are at least three different transitions that are happening when opening and closing the toast. The last one so far is for `opacity`.  This marks the end of the transition, so we check for this to determine if this is the correct event.